Develop branch tested with CC 10.3 against 10.2 release

+++ Jenkinsfile
@@ -1,56 +1,21 @@
 // curl -X POST -F "jenkinsfile=<Jenkinsfile" http://ccbvtauto.eur.ad.sag:8080/pipeline-model-converter/validate
 
-def testTemplates(templates) {
-    for (t in templates) {
-        // do we have custom docker-componse.yml ?
-        boolean customEnv = fileExists("templates/$t/docker-compose.yml")
-        
+def testTemplate(t, testProvision, buildImage, pushImage) {
+    dir ("templates/$t") {
         try {
-            // init the env
-            if (customEnv) {
-                dir("templates/$t") {
-                    sh "docker-compose -p $t run --name $t-env --rm -e CC_TEMPLATE=$t init"
-                }
-            }
-            
-            sh "docker-compose run --name $t --rm -e CC_TEMPLATE=$t test"
-            
-            junit "build/tests/**/*.xml"
-            archive "build/templates/${t}.zip"
-        }
-        catch (err) {
-            // capture logs
-            if (customEnv) {
-                dir("templates/$t") {
-                    sh "docker-compose -p $t logs"
-                }
-            }        
-        } finally {
-            // down the env
-            if (customEnv) {
-                dir("templates/$t") {
-                    sh "docker-compose -p $t down"
-                }
-            }        
-
-            // cleanup build/ files
-            // sh "docker-compose run --name $t --rm -e CC_TEMPLATE=$t test antcc clean"
-        }       
-    }   
-}
-
-def testTemplates2(templates) {
-    for (t in templates) {
-        dir ("templates/$t") {
-            try {
+            if (testProvision) {
                 sh "docker-compose run --name $t --rm provision"
-                sh "docker-compose ps"
-            } finally {
-                sh "docker-compose logs"
-                sh "docker-compose down"
             }
+            if (buildImage) {
+                sh "docker-compose build"
+                if (pushImage) {
+                    sh "docker-compose push"
+                }
+            }
+        } finally {
+            sh "docker-compose down"
         }
-    }   
+    }
 }
 
 pipeline {
@@ -58,54 +23,56 @@ pipeline {
         label 'docker'
     }
     environment {
-        RELEASE = '10.2'    // MUST match the release
-        TAG = "10.2.0.1.31" // All images tag      
-        
-        SAG_AQUARIUS = 'aquarius-bg.eur.ad.sag'
-        EMPOWER = credentials('empower')
+        TAG = "10.2"  // 10.3, 10.2, 10.1
+        FIXES = 'ALL' // for GA releases or '' for TRUNK
     }
     stages {
         stage('Init') {
             steps {
-                sh 'docker-compose up -d --build builder' // build and start the builder
-                //sh 'docker-compose up -d cc'
-                //sh 'docker-compose port cc 8091'
+                echo "Testing for ${env.TAG} release"
+                sh ". ./${env.TAG}.env; docker-compose pull cc"
             }
         }
-        stage("Test") {
+        stage("Level 1") {
             parallel {
-                // stage('Command Central') {
-                //     steps {
-                //         testTemplates(['sag-creds', 'sag-repos', 'sag-cc-tuneup'])
-                //     }
-                // }
-                // stage('Jenkins') {
-                //     steps {
-                //         testTemplates(['jenkins'])
-                //     }
-                // }
                 stage('Universal Messaging') {
                     steps {
-                        testTemplates2(['sag-um-server'])
+                        testTemplate('sag-um-server', false, true, true)
                     }
                 }
-                // stage('Terracotta') {
-                //     steps {
-                //         testTemplates([])
-                //     }
-                // }
+                stage('Terracotta') {
+                    steps {
+                        testTemplate('sag-tc-server', true, false, false)
+                    }
+                }
                 stage('Integration Server') {
                     steps {
-                        testTemplates2(['sag-msc-server'])
+                        testTemplate('sag-msc-server', false, true, true)
+                    }
+                }                                                
+            }
+        }
+        stage("Level 2") {
+            parallel {
+                stage('EntireX') {
+                    steps {
+                        testTemplate('sag-exx-broker', true, false, false)
+                    }
+                }
+                stage('Designer') {
+                    steps {
+                        testTemplate('sag-designer-services', true, false, false)
+                    }
+                }
+                stage('Apama') {
+                    steps {
+                        testTemplate('sag-apama-correlator', true, false, false)
                     }
                 }                                                
             }
         }
     }
     post {
-        failure {
-            sh 'docker-compose logs cc'
-        }
         always {
             sh 'docker-compose down'
         }
